+++ Manager/Abstract.php
@@ -1,10 +1,12 @@
 <?php
+namespace xiaogouxo\phpfetcher\Manager;
 /*
  * @author xuruiqi
  * @date   20150406
  * @copyright reetsee.com
- */     
-class Phpfetcher_Manager_Abstract {
+ */
+
+class Manager_Abstract {
 
 }
 ?>
