--- demo/structure
@@ -1,15 +1,0 @@
-|-phpfetcher.php                //使用时include这个文件即可
-|-Phpfetcher                    //此文件夹与phpfetcher.php要在同一目录下
-    |-Error.php                 //出错处理相关
-    |-Log.php                   //日志相关
-    |-Manager                   //暂时无用
-    |-Crawler                   //Crawler目录，存放爬虫相关的类代码
-        |-Abstract.php          //爬虫基类Phpfetcher_Crawler_Abstrct
-        |-Default.php           //默认提供的爬虫类Phpfetcher_Crawler_Default
-    |-Page                      //Page目录，存放页面相关的类代码
-        |-Abstract.php          //页面基类Phpfetcher_Page_Abstract
-        |-Default.php           //默认提供的页面类Phpfetcher_Page_Default
-    |-Dom                       //Dom目录，存放Dom相关的类代码
-        |-Abstract.php          //Dom基类Phpfetcher_Dom_Abstract
-        |-SimpleHtmlDom.php     //默认提供的Dom类Phpfetcher_Dom_SimpleHtmlDom
-        |-simple_html_dom.php   //Phpfetcher_Dom_SimpleHtmlDom实际使用了simple_html_dom中的代码
