--- tests/phpinfo_test.php
@@ -1,3 +1,0 @@
-<?php
-echo phpinfo();
-?>
